Add block_with_incorrect_hash for genesis-praos
