+++ testing/jormungandr-integration-tests/src/jormungandr/bft/block.rs
@@ -3,9 +3,7 @@ use chain_impl_mockchain::{
     block::{builder, BlockDate, BlockVersion, Contents, ContentsBuilder},
     chaintypes::{ConsensusType, ConsensusVersion},
     fee::LinearFee,
-    header::HeaderBuilder,
     key::BftLeaderId,
-    testing::TestGen,
 };
 use jormungandr_lib::interfaces::SlotDuration;
 use jormungandr_testing_utils::testing::{
@@ -14,47 +12,6 @@ use jormungandr_testing_utils::testing::{
     network::{builder::NetworkBuilder, Blockchain, Node, SpawnParams, Topology},
     startup, FragmentBuilder,
 };
-
-#[test]
-/// Ensures that blocks with an incorrect content hash are rejected by a BFT leader node
-fn block_with_incorrect_hash() {
-    let temp_dir = TempDir::new().unwrap();
-    let keys = startup::create_new_key_pair();
-
-    let node_params = ConfigurationBuilder::default()
-        .with_block0_consensus(ConsensusType::Bft)
-        .with_slot_duration(10)
-        .with_leader_key_pair(keys.clone())
-        .build(&temp_dir);
-
-    let block0 = node_params.block0_configuration().to_block();
-
-    let jormungandr = Starter::default().config(node_params).start().unwrap();
-
-    let contents = Contents::empty();
-    let content_size = contents.compute_hash_size().1;
-
-    let block = builder(BlockVersion::Ed25519Signed, contents, |_| {
-        Ok::<_, ()>({
-            HeaderBuilder::new_raw(BlockVersion::Ed25519Signed, &TestGen::hash(), content_size)
-                .set_parent(&block0.header().id(), 1.into())
-                .set_date(BlockDate {
-                    epoch: 0,
-                    slot_id: 1,
-                })
-                .into_bft_builder()
-                .unwrap()
-                .sign_using(keys.0.private_key())
-                .generalize()
-        })
-    })
-    .unwrap();
-
-    assert!(AdversaryNodeBuilder::new(block0)
-        .build()
-        .send_block_to_peer(jormungandr.address(), block)
-        .is_err());
-}
 
 #[test]
 /// Ensures that blocks with an incorrect signature are rejected by a BFT leader node
